修复主键重复bug

+++ Xin.ExternalService.EC/Response/Model/EC_ReceivingDetail.cs
@@ -161,7 +161,7 @@ namespace Xin.ExternalService.EC.Response.Model
         ///运费
         /// <summary> 
         [JsonProperty(PropertyName = "shipping_fee", NullValueHandling = NullValueHandling.Ignore)]
-        public int? ShippingFee { get; set; }
+        public decimal? ShippingFee { get; set; }
         /// <summary> 
         ///关税
         /// <summary>
